+++ src/hr/avrbanac/openglplayground/Globals.java
@@ -12,8 +12,10 @@ public class Globals {
     public static final int DISPLAY_HEIGHT          = 720;
     public static final String DISPLAY_TITLE        = "GLFW window";
     
-    public static final String VERTEX_FILE          = "src/hr/avrbanac/openglplayground/shaders/shader.vert";
-    public static final String FRAGMENT_FILE        = "src/hr/avrbanac/openglplayground/shaders/shader.frag";
+    public static final String VERTEX_FILE              = "src/hr/avrbanac/openglplayground/shaders/shader.vert";
+    public static final String FRAGMENT_FILE            = "src/hr/avrbanac/openglplayground/shaders/shader.frag";
+    public static final String TERRAIN_VERTEX_FILE      = "src/hr/avrbanac/openglplayground/shaders/terrainShader.vert";
+    public static final String TERRAIN_FRAGMENT_FILE    = "src/hr/avrbanac/openglplayground/shaders/terrainShader.frag";
     
     public static final String TEXTURE_TYPE         = "PNG";
     public static final String TEXTURE_FOLDER       = "res/";
@@ -21,8 +23,11 @@ public class Globals {
     public static final String MODEL_FOLDER         = "res/";
     public static final String MODEL_EXTENSION      = ".obj";
     
-    public static final float FOV                   = 70;
+    public static final float FOV                   = 70f;
     public static final float NEAR_PLANE            = 0.1f;
     public static final float FAR_PLANE             = 1000.0f;
+    
+    public static final float TERRAIN_SIZE          = 800;
+    public static final int TERRAIN_VERTEX_COUNT    = 128;
 }